fix remove workflow catalogue bug

+++ dlink-admin/src/main/java/com/dlink/service/impl/WorkflowCatalogueServiceImpl.java
@@ -2,6 +2,7 @@ package com.dlink.service.impl;
 
 import cn.hutool.json.JSON;
 import cn.hutool.json.JSONObject;
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.dlink.db.service.impl.SuperServiceImpl;
 import com.dlink.dto.CatalogueTaskDTO;
@@ -14,6 +15,7 @@ import com.dlink.scheduler.client.ProcessClient;
 import com.dlink.scheduler.client.ProjectClient;
 import com.dlink.scheduler.model.ProcessDefinition;
 import com.dlink.scheduler.model.Project;
+import com.dlink.scheduler.result.Result;
 import com.dlink.service.JobInstanceService;
 import com.dlink.service.WorkflowCatalogueService;
 import com.dlink.service.WorkflowTaskService;
@@ -48,6 +50,8 @@ public class WorkflowCatalogueServiceImpl extends SuperServiceImpl<WorkflowCatal
     private ProcessClient processClient;
     @Autowired
     private ProjectClient projectClient;
+    @Autowired
+    private WorkflowCatalogueService catalogueService;
 
     @Override
     public List<WorkflowCatalogue> getAllData() {
@@ -63,7 +67,7 @@ public class WorkflowCatalogueServiceImpl extends SuperServiceImpl<WorkflowCatal
     @Override
     public boolean createCatalogue(WorkflowCatalogue catalogue) {
         // 创建dolphinscheduler的项目, 只有根目录才会去创建项目
-        if (catalogue.getParentId() == null) {
+        if (catalogue.getParentId() == null || catalogue.getParentId() == 0) {
             Project project = null;
             if (StringUtils.isBlank(catalogue.getProjectCode())) {
                 project = projectClient.createProjectByName(catalogue.getName());
@@ -120,26 +124,22 @@ public class WorkflowCatalogueServiceImpl extends SuperServiceImpl<WorkflowCatal
         } else {
             if (isNotNull(catalogue.getTaskId())) {
                 Integer taskId = catalogue.getTaskId();
-                JobInstance job = jobInstanceService.getJobInstanceByTaskId(taskId);
-                if (job == null
-                        || (JobStatus.FINISHED.getValue().equals(job.getStatus())
-                        || JobStatus.FAILED.getValue().equals(job.getStatus())
-                        || JobStatus.CANCELED.getValue().equals(job.getStatus())
-                        || JobStatus.UNKNOWN.getValue().equals(job.getStatus()))) {
-                    // 删除作业同时将海豚调度上线的任务删除
+                WorkflowTaskDTO taskInfo = taskService.getTaskInfoById(taskId);
+                if (taskInfo.getStatus().equals(WorkflowLifeCycle.ONLINE.getValue())) {
+                    errors.add(taskInfo.getName());
+                } else {
+                    // 删除作业同时将海豚调度上的任务删除
                     this.removeWorkflowInDSByTaskId(catalogue.getTaskId());
                     taskService.removeById(taskId);
                     if (StringUtils.isNotBlank(catalogue.getProjectCode())) {
                         // 删除dolphinscheduler的项目
-                        JSONObject result = projectClient.deleteProjectByName(catalogue.getProjectCode());
-                        if (result.getBool("success")) {
+                        Result<JSONObject> result = projectClient.deleteProjectByCode(catalogue.getProjectCode());
+                        if (result.getSuccess()) {
                             this.removeById(id);
                         }
                     } else {
                         this.removeById(id);
                     }
-                } else {
-                    errors.add(job.getName());
                 }
             } else {
                 List<WorkflowCatalogue> all = this.getAllData();
@@ -148,9 +148,21 @@ public class WorkflowCatalogueServiceImpl extends SuperServiceImpl<WorkflowCatal
                 List<String> actives = this.analysisActiveCatalogues(del);
                 if (actives.isEmpty()) {
                     for (WorkflowCatalogue c : del) {
-                        // 删除作业同时将海豚调度上线的任务删除
-                        this.removeWorkflowInDSByTaskId(c.getTaskId());
-                        taskService.removeById(c.getTaskId());
+                        // 删除作业同时将海豚调度上的任务删除
+                        if (c.getTaskId() != null) {
+                            this.removeWorkflowInDSByTaskId(c.getTaskId());
+                            taskService.removeById(c.getTaskId());
+                        }
+
+                        if (StringUtils.isNotBlank(catalogue.getProjectCode())) {
+                            // 删除dolphinscheduler的项目
+                            Result<JSONObject> result = projectClient.deleteProjectByCode(catalogue.getProjectCode());
+                            if (result.getSuccess()) {
+                                this.removeById(id);
+                            }
+                        } else {
+                            this.removeById(id);
+                        }
                         this.removeById(c.getId());
                     }
                 } else {
@@ -162,15 +174,25 @@ public class WorkflowCatalogueServiceImpl extends SuperServiceImpl<WorkflowCatal
         return errors;
     }
 
+    private WorkflowCatalogue getRootCatalogueByCatalogue(WorkflowCatalogue catalogue) {
+        if (catalogue.getParentId() != null && catalogue.getParentId() > 0) {
+            catalogue = catalogueService.getById(catalogue.getParentId());
+            return getRootCatalogueByCatalogue(catalogue);
+        }
+        return catalogue;
+    }
+
     private void removeWorkflowInDSByTaskId(Integer taskId) {
         WorkflowTaskDTO taskInfo = taskService.getTaskInfoById(taskId);
         if (taskInfo != null) {
-            Project dinkyProject = SystemInit.getProject();
-            long projectCode = dinkyProject.getCode();
+            // 获取根目录
+            WorkflowCatalogue catalogue = catalogueService.getOne(new LambdaQueryWrapper<WorkflowCatalogue>().eq(WorkflowCatalogue::getTaskId, taskId));
+            WorkflowCatalogue root = getRootCatalogueByCatalogue(catalogue);
+            long projectCode = Long.valueOf(root.getProjectCode());
             ProcessDefinition process = processClient.getProcessDefinitionInfo(projectCode, taskInfo.getName());
             if (process != null) {
                 // 先下线
-                processClient.onlineProcessDefinition(projectCode, process, "OFFLINE");
+//                processClient.onlineProcessDefinition(projectCode, process, "OFFLINE");
                 // 再删除
                 processClient.deleteProcessDefinition(projectCode, process.getCode());
             }
@@ -178,13 +200,17 @@ public class WorkflowCatalogueServiceImpl extends SuperServiceImpl<WorkflowCatal
     }
 
     private List<String> analysisActiveCatalogues(Set<WorkflowCatalogue> del) {
-        List<Integer> actives = jobInstanceService.listJobInstanceActive().stream().map(JobInstance::getTaskId)
-                .collect(Collectors.toList());
         List<WorkflowCatalogue> activeCatalogue = del.stream()
-                .filter(catalogue -> catalogue.getTaskId() != null && actives.contains(catalogue.getTaskId()))
+                .filter(catalogue -> catalogue.getTaskId() != null)
                 .collect(Collectors.toList());
-        return activeCatalogue.stream().map(catalogue -> taskService.getById(catalogue.getTaskId()).getName())
-                .collect(Collectors.toList());
+        List<String> result = new ArrayList<>();
+        for (WorkflowCatalogue catalogue : activeCatalogue) {
+            WorkflowTask workflowTask = taskService.getById(catalogue.getTaskId());
+            if (workflowTask.getStatus().equals(WorkflowLifeCycle.ONLINE)) {
+                result.add(workflowTask.getName());
+            }
+        }
+        return result;
     }
 
     private void findAllCatalogueInDir(Integer id, List<WorkflowCatalogue> all, Set<WorkflowCatalogue> del) {
